Refactor data index to use Pointers struct
Refactor the data index to use a new Pointers struct instead of using a
vec. This allow for better encapsulation of the pointers, allowing
common methods to be implemented on the struct. This also allows for the
potential of changing out the internal data structure in the future if
desired. The data leaf will be changed to use Pointers in a future
commit.